+++ pajbot/modules/__init__.py
@@ -18,7 +18,6 @@ from pajbot.modules.basic.ignore import IgnoreModule
 from pajbot.modules.basic.permaban import PermabanModule
 from pajbot.modules.basic.pointsreset import PointsResetModule
 from pajbot.modules.basic.stream_update import StreamUpdateModule
-from pajbot.modules.basic.twitter import TwitterModule
 from pajbot.modules.bingo import BingoModule
 from pajbot.modules.casechecker import CaseCheckerModule
 from pajbot.modules.chat_alerts import ChatAlertModule
@@ -149,7 +148,6 @@ available_modules = [
     SubscriberFetchModule,
     TopModule,
     TriviaModule,
-    TwitterModule,
     TypeEmoteQuestModule,
     TypeMeMessageQuestModule,
     VanishModule,
